test for using VDS as priority sites passes
also noticed that "limit" was hard coded into the groupsites code.
switched it for distance.

also noticed that I'm not really checking to see whether or not a
detector is farther away from others right now.  That is, I am using
"distance" in assigning the priority sites, but if two priority sites
are less than "distance" apart, then one of them might not be used.